+++ moabb/evaluations/evaluations.py
@@ -12,14 +12,21 @@ from sklearn.model_selection import (
     LeaveOneGroupOut,
     StratifiedKFold,
     StratifiedShuffleSplit,
-    cross_validate,
 )
-from sklearn.model_selection._validation import _fit_and_score, _score
+from sklearn.model_selection._validation import _score
 from sklearn.preprocessing import LabelEncoder
 from tqdm import tqdm
 
 from moabb.evaluations.base import BaseEvaluation
-from moabb.evaluations.utils import create_save_path, save_model_cv, save_model_list
+from moabb.evaluations.splitters import (
+    CrossSessionSplitter,
+    CrossSubjectSplitter,
+    WithinSessionSplitter,
+)
+from moabb.evaluations.utils import (
+    _create_save_path,
+    _save_model_cv,
+)
 
 
 try:
@@ -28,6 +35,7 @@ try:
     _carbonfootprint = True
 except ImportError:
     _carbonfootprint = False
+
 
 log = logging.getLogger(__name__)
 
@@ -134,7 +142,6 @@ class WithinSessionEvaluation(BaseEvaluation):
             super().__init__(**kwargs)
 
     # flake8: noqa: C901
-
     def _evaluate(
         self,
         dataset,
@@ -172,8 +179,13 @@ class WithinSessionEvaluation(BaseEvaluation):
                         # Initialize CodeCarbon
                         tracker = EmissionsTracker(save_to_file=False, log_level="error")
                         tracker.start()
+
                     t_start = time()
-                    cv = StratifiedKFold(5, shuffle=True, random_state=self.random_state)
+                    self.cv = WithinSessionSplitter(
+                        n_folds=5,
+                        shuffle=True,
+                        random_state=self.random_state,
+                    )
                     inner_cv = StratifiedKFold(
                         3, shuffle=True, random_state=self.random_state
                     )
@@ -185,17 +197,6 @@ class WithinSessionEvaluation(BaseEvaluation):
 
                     grid_clf = clone(clf)
 
-                    # Create folder for grid search results
-                    create_save_path(
-                        self.hdf5_path,
-                        dataset.code,
-                        subject,
-                        session,
-                        name,
-                        grid=True,
-                        eval_type="WithinSession",
-                    )
-
                     # Implement Grid Search
                     grid_clf = self._grid_search(
                         param_grid=param_grid,
@@ -203,64 +204,51 @@ class WithinSessionEvaluation(BaseEvaluation):
                         grid_clf=grid_clf,
                         inner_cv=inner_cv,
                     )
+
                     if self.hdf5_path is not None and self.save_model:
-                        model_save_path = create_save_path(
+                        model_save_path = _create_save_path(
                             self.hdf5_path,
                             dataset.code,
                             subject,
                             session,
                             name,
-                            grid=False,
+                            grid=self.search,
                             eval_type="WithinSession",
                         )
 
-                    if isinstance(X, BaseEpochs):
-                        scorer = get_scorer(self.paradigm.scoring)
-                        acc = list()
-                        X_ = X[ix]
-                        y_ = y[ix] if self.mne_labels else y_cv
-                        for cv_ind, (train, test) in enumerate(cv.split(X_, y_)):
-                            cvclf = clone(grid_clf)
-                            cvclf.fit(X_[train], y_[train])
-                            acc.append(scorer(cvclf, X_[test], y_[test]))
+                    scorer = get_scorer(self.paradigm.scoring)
+                    acc = list()
+                    X_ = X[ix]
+                    y_ = y[ix] if self.mne_labels else y_cv
+                    meta_ = metadata[ix].reset_index(drop=True)
 
-                            if self.hdf5_path is not None and self.save_model:
-                                save_model_cv(
-                                    model=cvclf,
-                                    save_path=model_save_path,
-                                    cv_index=cv_ind,
-                                )
+                    for cv_ind, (train, test) in enumerate(self.cv.split(y_, meta_)):
+                        cvclf = clone(grid_clf)
 
-                        acc = np.array(acc)
-                        score = acc.mean()
-                    else:
-                        results = cross_validate(
-                            grid_clf,
-                            X[ix],
-                            y_cv,
-                            cv=cv,
-                            scoring=self.paradigm.scoring,
-                            n_jobs=self.n_jobs,
-                            error_score=self.error_score,
-                            return_estimator=True,
-                        )
-                        score = results["test_score"].mean()
+                        cvclf.fit(X_[train], y_[train])
+
+                        score = scorer(cvclf, X_[test], y_[test])
+
+                        acc.append(score)
+
                         if self.hdf5_path is not None and self.save_model:
-                            save_model_list(
-                                results["estimator"],
-                                score_list=results["test_score"],
+                            _save_model_cv(
+                                model=cvclf,
                                 save_path=model_save_path,
+                                cv_index=cv_ind,
                             )
+
+                    acc = np.array(acc)
+                    score = acc.mean()
 
                     if _carbonfootprint:
                         emissions = tracker.stop()
                         if emissions is None:
                             emissions = np.nan
                     duration = time() - t_start
-
                     nchan = X.info["nchan"] if isinstance(X, BaseEpochs) else X.shape[1]
                     res = {
-                        "time": duration / 5.0,  # 5 fold CV
+                        "time": duration / self.cv.n_folds,  # 5 fold CV
                         "dataset": dataset,
                         "subject": subject,
                         "session": session,
@@ -519,7 +507,8 @@ class CrossSessionEvaluation(BaseEvaluation):
                     tracker.start()
 
                 # we want to store a results per session
-                cv = LeaveOneGroupOut()
+                self.cv = CrossSessionSplitter(random_state=self.random_state)
+
                 inner_cv = StratifiedKFold(
                     3, shuffle=True, random_state=self.random_state
                 )
@@ -532,62 +521,42 @@ class CrossSessionEvaluation(BaseEvaluation):
                 )
 
                 if self.hdf5_path is not None and self.save_model:
-                    model_save_path = create_save_path(
+                    model_save_path = _create_save_path(
                         hdf5_path=self.hdf5_path,
                         code=dataset.code,
                         subject=subject,
                         session="",
                         name=name,
-                        grid=False,
+                        grid=self.search,
                         eval_type="CrossSession",
                     )
 
-                for cv_ind, (train, test) in enumerate(cv.split(X, y, groups)):
+                for cv_ind, (train, test) in enumerate(self.cv.split(y, metadata)):
                     model_list = []
                     if _carbonfootprint:
                         tracker.start()
                     t_start = time()
-                    if isinstance(X, BaseEpochs):
-                        cvclf = clone(grid_clf)
-                        cvclf.fit(X[train], y[train])
-                        model_list.append(cvclf)
-                        score = scorer(cvclf, X[test], y[test])
 
-                        if self.hdf5_path is not None and self.save_model:
-                            save_model_cv(
-                                model=cvclf,
-                                save_path=model_save_path,
-                                cv_index=str(cv_ind),
-                            )
-                    else:
-                        result = _fit_and_score(
-                            estimator=clone(grid_clf),
-                            X=X,
-                            y=y,
-                            scorer=scorer,
-                            train=train,
-                            test=test,
-                            verbose=False,
-                            parameters=None,
-                            fit_params=None,
-                            error_score=self.error_score,
-                            return_estimator=True,
-                            score_params={},
+                    cvclf = clone(grid_clf)
+
+                    cvclf.fit(X[train], y[train])
+
+                    model_list.append(cvclf)
+                    score = scorer(cvclf, X[test], y[test])
+
+                    if self.hdf5_path is not None and self.save_model:
+                        _save_model_cv(
+                            model=cvclf,
+                            save_path=model_save_path,
+                            cv_index=str(cv_ind),
                         )
-                        score = result["test_scores"]
-                        model_list = result["estimator"]
+
                     if _carbonfootprint:
                         emissions = tracker.stop()
                         if emissions is None:
                             emissions = 0
 
                     duration = time() - t_start
-                    if self.hdf5_path is not None and self.save_model:
-                        save_model_list(
-                            model_list=model_list,
-                            score_list=score,
-                            save_path=model_save_path,
-                        )
 
                     nchan = X.info["nchan"] if isinstance(X, BaseEpochs) else X.shape[1]
                     res = {
@@ -677,7 +646,6 @@ class CrossSubjectEvaluation(BaseEvaluation):
         if len(run_pipes) == 0:
             return
 
-        # get the data
         X, y, metadata = self.paradigm.get_data(
             dataset=dataset,
             return_epochs=self.return_epochs,
@@ -686,8 +654,6 @@ class CrossSubjectEvaluation(BaseEvaluation):
             postprocess_pipeline=postprocess_pipeline,
             process_pipelines=[process_pipeline],
         )
-
-        # encode labels
         le = LabelEncoder()
         y = y if self.mne_labels else le.fit_transform(y)
 
@@ -700,15 +666,20 @@ class CrossSubjectEvaluation(BaseEvaluation):
 
         # perform leave one subject out CV
         if self.n_splits is None:
-            cv = LeaveOneGroupOut()
+            cv_class = LeaveOneGroupOut
+            cv_kwargs = {}
         else:
-            cv = GroupKFold(n_splits=self.n_splits)
+            cv_class = GroupKFold
+            cv_kwargs = {"n_splits": self.n_splits}
             n_subjects = self.n_splits
+
+        self.cv = CrossSubjectSplitter(
+            cv_class=cv_class, random_state=self.random_state, **cv_kwargs
+        )
 
         inner_cv = StratifiedKFold(3, shuffle=True, random_state=self.random_state)
 
         # Implement Grid Search
-
         if _carbonfootprint:
             # Initialise CodeCarbon
             tracker = EmissionsTracker(save_to_file=False, log_level="error")
@@ -716,7 +687,7 @@ class CrossSubjectEvaluation(BaseEvaluation):
         # Progressbar at subject level
         for cv_ind, (train, test) in enumerate(
             tqdm(
-                cv.split(X, y, groups),
+                self.cv.split(y, metadata),
                 total=n_subjects,
                 desc=f"{dataset.code}-CrossSubject",
             )
@@ -734,7 +705,26 @@ class CrossSubjectEvaluation(BaseEvaluation):
                 clf = self._grid_search(
                     param_grid=param_grid, name=name, grid_clf=clf, inner_cv=inner_cv
                 )
+
+                if self.hdf5_path is not None and self.save_model:
+                    # Save the best model from grid search
+                    model_save_path = _create_save_path(
+                        hdf5_path=self.hdf5_path,
+                        code=dataset.code,
+                        subject=subject,
+                        session="",
+                        name=name,
+                        grid=self.search,
+                        eval_type="CrossSubject",
+                    )
+                    _save_model_cv(
+                        model=clf,
+                        save_path=model_save_path,
+                        cv_index=str(cv_ind),
+                    )
+
                 model = deepcopy(clf).fit(X[train], y[train])
+
                 if _carbonfootprint:
                     emissions = tracker.stop()
                     if emissions is None:
@@ -742,17 +732,18 @@ class CrossSubjectEvaluation(BaseEvaluation):
                 duration = time() - t_start
 
                 if self.hdf5_path is not None and self.save_model:
-                    model_save_path = create_save_path(
+
+                    model_save_path = _create_save_path(
                         hdf5_path=self.hdf5_path,
                         code=dataset.code,
                         subject=subject,
                         session="",
                         name=name,
-                        grid=False,
+                        grid=self.search,
                         eval_type="CrossSubject",
                     )
 
-                    save_model_cv(
+                    _save_model_cv(
                         model=model, save_path=model_save_path, cv_index=str(cv_ind)
                     )
                 # we eval on each session